Adding Test users data
- Adding more test to users models views
- Giving meaningful names to existing test functions
- Moving the users, posts, direct_message tests to global folder.
- Changing the tests file accordingly
