+++ tests/direct_message_tests/tests.py
@@ -1,30 +1,7 @@
-from . import models
-import pytest
-from django.utils import timezone
+from direct_message.models import Message
 from django.contrib.auth.models import User
-
-
-CREATION_DATE = timezone.now()
-
-
-@pytest.fixture
-def new_user_sender():
-    return User(username='TestSender', first_name='Nick',
-                last_name='Birch', password='1234', email='Nick@email.com')
-
-
-@pytest.fixture
-def new_user_receiver():
-    return User(username='TestReceiver', first_name='Andrew',
-                last_name='Glouberman', password='abcd', email='Andrew@email.com')
-
-
-@pytest.fixture
-def new_message(new_user_sender, new_user_receiver):
-    return models.Message(sender=new_user_sender,
-                          receiver=new_user_receiver,
-                          creation_date=CREATION_DATE,
-                          message_text='test')
+import pytest
+from conftest import CREATION_DATE
 
 
 def save_a_message(new_message):
@@ -56,27 +33,27 @@ def test_is_empty_message(new_message, empty_string):
 @pytest.mark.django_db()
 def test_saving_message_in_db(new_message):
     save_a_message(new_message)
-    assert new_message.sender in models.User.objects.all()
-    assert new_message.receiver in models.User.objects.all()
-    assert new_message in models.Message.objects.all()
+    assert new_message.sender in User.objects.all()
+    assert new_message.receiver in User.objects.all()
+    assert new_message in Message.objects.all()
 
 
 @pytest.mark.django_db()
 def test_deletion_message_from_db(new_message):
     save_a_message(new_message)
     new_message.delete()
-    assert new_message not in models.Message.objects.all()
+    assert new_message not in Message.objects.all()
 
 
 @pytest.mark.django_db()
 def test_deletion_message_after_sender_deletion(new_message):
     save_a_message(new_message)
     new_message.sender.delete()
-    assert new_message not in models.Message.objects.all()
+    assert new_message not in Message.objects.all()
 
 
 @pytest.mark.django_db()
 def test_deletion_message_after_receiver_deletion(new_message):
     save_a_message(new_message)
     new_message.receiver.delete()
-    assert new_message not in models.Message.objects.all()
+    assert new_message not in Message.objects.all()
